@@ -1,10 +1,11 @@
+# frozen_string_literal: true
 module Fogged
   def self.directory_public_url(directory_name)
     case Fogged.provider
     when :aws
-      Fogged.storage.request_url(:bucket_name => directory_name)
+      Fogged.storage.request_url(bucket_name: directory_name)
     else
-      fail(ArgumentError, "Provider #{Fogged.provider} is not available!")
+      raise(ArgumentError, "Provider #{Fogged.provider} is not available!")
     end
   end
 
@@ -15,8 +16,8 @@ module Fogged
   def self.file_public_url(key, directory = Fogged.resources.key)
     Fogged.storage.try(
       :request_url,
-      :bucket_name => directory,
-      :object_name => key
+      bucket_name: directory,
+      object_name: key
     )
   end
 
