@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "test_helper"
 
 module Fogged
@@ -7,16 +8,16 @@ module Fogged
 
     setup do
       @resource = fogged_resources(:resource_mov_3)
-      @resource.update!(:encoding_job_id => 1234565434567890)
+      @resource.update!(encoding_job_id: 1_234_565_434_567_890)
     end
 
     test "should receive zencoder notification" do
-      refute @resource.encoding_progress
-      refute @resource.width
-      refute @resource.height
-      refute @resource.duration
+      assert_not @resource.encoding_progress
+      assert_not @resource.width
+      assert_not @resource.height
+      assert_not @resource.duration
 
-      post :zencoder_notification, :params => payload(@resource.encoding_job_id)
+      post :zencoder_notification, params: payload(@resource.encoding_job_id)
 
       assert @resource.reload.encoding_progress
       assert @resource.width
@@ -25,49 +26,49 @@ module Fogged
     end
 
     test "should receive zencoder notification with invalid id" do
-      post :zencoder_notification, :params => payload("foobar!")
+      post :zencoder_notification, params: payload("foobar!")
     end
 
     test "should receive zencoder notification with invalid payload" do
-      refute @resource.encoding_progress
-      refute @resource.width
-      refute @resource.height
-      refute @resource.duration
+      assert_not @resource.encoding_progress
+      assert_not @resource.width
+      assert_not @resource.height
+      assert_not @resource.duration
 
       post :zencoder_notification,
-           :params => payload(@resource.encoding_job_id).except(:outputs)
+           params: payload(@resource.encoding_job_id).except(:outputs)
 
-      refute @resource.reload.encoding_progress
-      refute @resource.width
-      refute @resource.height
-      refute @resource.duration
+      assert_not @resource.reload.encoding_progress
+      assert_not @resource.width
+      assert_not @resource.height
+      assert_not @resource.duration
     end
 
     test "should receive zencoder notification with failed job" do
-      refute @resource.encoding_progress
-      refute @resource.width
-      refute @resource.height
-      refute @resource.duration
+      assert_not @resource.encoding_progress
+      assert_not @resource.width
+      assert_not @resource.height
+      assert_not @resource.duration
 
       post :zencoder_notification,
-           :params => payload(@resource.encoding_job_id, "failed")
+           params: payload(@resource.encoding_job_id, "failed")
 
-      refute @resource.reload.encoding_progress
-      refute @resource.width
-      refute @resource.height
-      refute @resource.duration
+      assert_not @resource.reload.encoding_progress
+      assert_not @resource.width
+      assert_not @resource.height
+      assert_not @resource.duration
     end
 
     private
 
     def payload(encoding_job_id, state = "finished")
       {
-        :job => {
-          :id => encoding_job_id,
-          :state => state
+        job: {
+          id: encoding_job_id,
+          state:
         },
-        :outputs => [
-          { :duration_in_ms => 85600, :height => 480, :width => 640 }
+        outputs: [
+          { duration_in_ms: 85_600, height: 480, width: 640 }
         ]
       }
     end
