@@ -1,5 +1,6 @@
+# frozen_string_literal: true
 class ImagesControllers < ApplicationController
   def index
-    render :json => Image.all
+    render json: Image.all
   end
 end
