@@ -1,3 +1,4 @@
-class Image < ActiveRecord::Base
+# frozen_string_literal: true
+class Image < ApplicationRecord
   has_one_resource
 end
