@@ -1,4 +1,5 @@
-class MovieFoggedResource < ActiveRecord::Base
+# frozen_string_literal: true
+class MovieFoggedResource < ApplicationRecord
   belongs_to :movie
-  belongs_to :resource, :class_name => "Fogged::Resource"
+  belongs_to :resource, class_name: "Fogged::Resource"
 end
