@@ -1,4 +1,5 @@
+# frozen_string_literal: true
 Rails.application.routes.draw do
-  resources :images, :only => :index
+  resources :images, only: :index
   mount Fogged::Engine => "/"
 end
