@@ -1,6 +1,7 @@
+# frozen_string_literal: true
 class CreateMovies < ActiveRecord::Migration
   def change
-    create_table :movies, :force => true do |t|
+    create_table :movies, force: true do |t|
       t.string :name
 
       t.timestamps
