@@ -1,9 +1,10 @@
+# frozen_string_literal: true
 class CreateDelayedJobs < ActiveRecord::Migration
   def self.up
-    create_table :delayed_jobs, :force => true do |table|
-      table.integer :priority, :default => 0, :null => false # Allows some jobs to jump to the front of the queue
-      table.integer :attempts, :default => 0, :null => false # Provides for retries, but still fail eventually.
-      table.text :handler, :null => false                    # YAML-encoded string of the object that will do work
+    create_table :delayed_jobs, force: true do |table|
+      table.integer :priority, default: 0, null: false # Allows some jobs to jump to the front of the queue
+      table.integer :attempts, default: 0, null: false # Provides for retries, but still fail eventually.
+      table.text :handler, null: false # YAML-encoded string of the object that will do work
       table.text :last_error                                 # reason for last failure (See Note below)
       table.datetime :run_at                                 # When to run. Could be Time.zone.now for immediately, or sometime in the future.
       table.datetime :locked_at                              # Set when a client is working on this object
@@ -13,7 +14,7 @@ class CreateDelayedJobs < ActiveRecord::Migration
       table.timestamps
     end
 
-    add_index :delayed_jobs, [:priority, :run_at], :name => 'delayed_jobs_priority'
+    add_index :delayed_jobs, %i[priority run_at], name: "delayed_jobs_priority"
   end
 
   def self.down
