@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require "test_helper"
 
 class FoggedTest < ActiveSupport::TestCase
@@ -14,12 +15,12 @@ class FoggedTest < ActiveSupport::TestCase
   end
 
   test "should check if file exists" do
-    refute Fogged.file_exists?("foobar")
+    assert_not Fogged.file_exists?("foobar")
 
     Fogged.resources.files.create(
-      :key => "foobar",
-      :body => "foo",
-      :public => true
+      key: "foobar",
+      body: "foo",
+      public: true
     )
 
     assert Fogged.file_exists?("foobar")
