@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module ResourceTestHelper
   extend ActiveSupport::Concern
   include JsonTestHelper
@@ -19,16 +20,15 @@ module ResourceTestHelper
     json = response_json[json_key]
     assert_json_objects(resources, json, :id, :url, :name)
     resources.zip([json].flatten).each do |resource, json_resource|
-      if json_resource[:upload_url]
-        assert_equal resource.upload_url.chop, json_resource[:upload_url].chop
-      end
+      assert_equal resource.upload_url.chop, json_resource[:upload_url].chop if json_resource[:upload_url]
 
       next unless resource.video?
-      [:h264_url, :mpeg_url, :webm_url, :thumbnail_urls].each do |field|
+
+      %i[h264_url mpeg_url webm_url thumbnail_urls].each do |field|
         if resource.send(field)
           assert_equal resource.send(field), json_resource[field]
         else
-          refute json_resource[field]
+          assert_not json_resource[field]
         end
       end
       assert_equal((resource.encoding_progress || 0), json_resource[:encoding_progress])
