@@ -1,6 +1,7 @@
-require File.expand_path('../boot', __FILE__)
+# frozen_string_literal: true
+require File.expand_path("boot", __dir__)
 
-require 'rails/all'
+require "rails/all"
 
 Bundler.require(*Rails.groups)
 require "fogged"
@@ -20,6 +21,6 @@ module Dummy
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
-    config.autoload_paths += %W(#{Rails.root}/../controllers/concerns) if Rails.env.test?
+    config.autoload_paths += %W[#{Rails.root.join('../controllers/concerns')}] if Rails.env.test?
   end
 end
