@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 Rails.application.configure do
   config.active_support.test_order = :sorted
   # Settings specified here will take precedence over those in config/application.rb.
@@ -14,7 +15,7 @@ Rails.application.configure do
   config.eager_load = false
 
   # Configure static asset server for tests with Cache-Control for performance.
-  config.public_file_server.enabled  = true
+  config.public_file_server.enabled = true
   config.public_file_server.headers = { "Cache-Control" => "public, max-age=3600" }
 
   # Show full error reports and disable caching.
